小红花新增进度奖励领取（proto/IPC/活动页）；工具链升级 Rust 1.98.1：固定 rust-toolchain.toml、全库适配新 API 与 rustfmt、CI release 前置 fmt/clippy/test/前端检查 job；前端依赖批量升级（vue/pinia/naive-ui 等）

+++ crates/qq-farm-core/src/runtime/worker_loop/status.rs
@@ -40,7 +40,7 @@ impl WorkerLoop {
         let now = now_ms();
         let next = self.next_runs.lock().clone();
         // 量化到 5s 桶：否则倒计时字段每 3s 必变，门控永远不命中
-        let quantize = |ms: i64| (((ms / 1000).max(0) / 5) * 5) as i64;
+        let quantize = |ms: i64| ((ms / 1000).max(0) / 5) * 5;
         let farm = quantize(next.farm_at - now);
         let friend = quantize(next.friend_at - now);
         // help/steal 已合并为统一好友 tick（friend_at）；
